+++ cypress/support/index.js
@@ -1,4 +1,5 @@
-import '@bigtest/cypress';
+import { bigtestGlobals } from '@bigtest/globals';
+import '@interactors/with-cypress';
 
 import './stripes';
 
@@ -9,3 +10,6 @@ import './checkin';
 import './checkout';
 import './eholdings';
 import './inventory';
+import './users';
+
+bigtestGlobals.defaultInteractorTimeout = 10000;