+++ interactors/baseHTML.js
@@ -1,0 +1,6 @@
+import { HTML as BaseHTML } from '@interactors/html';
+
+export default BaseHTML
+  .filters({
+    element: element => element
+  });