+++ interactors/button.js
@@ -1,4 +1,4 @@
-import { HTML } from '@bigtest/interactor';
+import HTML from './baseHTML';
 
 export default HTML.extend('button')
   .selector('a[href],button,input[type=button],input[type=submit],input[type=reset],input[type=image]')
@@ -8,4 +8,11 @@ export default HTML.extend('button')
     anchor: (el) => el.tagName === 'A',
     default: (el) => el.classList.contains('default'),
     ariaLabel: (el) => el.ariaLabel,
+    disabled: {
+      apply: (el) => {
+        if (el.disabled !== undefined) return el.disabled;
+        return el.getAttribute('aria-disabled') === 'true';
+      },
+      default: false
+    }
   });