+++ interactors/pane.js
@@ -1,8 +1,12 @@
-import { HTML } from '@bigtest/interactor';
 import { isVisible } from 'element-is-visible';
 import Button from './button';
+import HTML from './baseHTML';
 
 function title(el) { return el.querySelector('[class^=paneTitle]').textContent; }
+
+export const PaneHeader = HTML.extend('pane header')
+  .selector('[class^=paneHeader-]')
+  .locator(title);
 
 export default HTML.extend('pane')
   .selector('[class^=pane-]')