+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Change history for stripes-testing
 
-## 3.1.0 (IN PROGRESS)
+## 4.1.0 (IN PROGRESS)
 
 * Export mocks for consumption in functional tests by apps. Refs UITEST-81.
+
+## [4.0.0](https://github.com/folio-org/stripes-testing/tree/v4.0.0) (2021-09-27)
+[Full Changelog](https://github.com/folio-org/stripes-testing/compare/v3.0.0...v4.0.0)
+
+* Introduce Cypress e2e tests with new BigTest. FOLIO-2946.
+* Introduce e2e tests with Cypress. FAT-899.
+* Write e2e tests with interactors.
 
 ## [3.0.0](https://github.com/folio-org/stripes-testing/tree/v3.0.0) (2021-03-18)
 [Full Changelog](https://github.com/folio-org/stripes-testing/compare/v2.0.1...v3.0.0)